requirements.txt
#requirements.txt
streamlit
requests
pandas
pytz
sqlalchemy
psycopg[binary,pool]   
psycopg2-binary   
apscheduler
python-dotenv
streamlit-card
streamlit-autorefresh  
streamlit-js-eval          
psutil 
aiohttp  # Added for async in sync.py